requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic
agno
anthropic
google-auth
google-api-python-client
